requirements.txt
Flask==3.0.0
pycryptodome==3.19.0
scikit-learn==1.3.0